username/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5



